chore(mdx): adjusting components, SEO, RSS
