+++ src/lib/mdx.ts
@@ -3,6 +3,7 @@ import fs from 'fs'
 import matter from 'gray-matter'
 import path from 'path'
 import renderToString from 'next-mdx-remote/render-to-string'
+import RSS from '@/scripts/generate-rss'
 
 const postDir = path.join(process.cwd(), 'posts')
 
@@ -29,12 +30,20 @@ export async function getFileBySlug(locale = 'id', slug = '') {
   return { mdxSource, ...data, slug }
 }
 
-export async function getAllFilesFrontMatter({ locale = 'id' }) {
+export async function getAllFilesFrontMatter({ locale = 'id', sort = true }) {
   const files = fs.readdirSync(path.join(postDir, locale))
 
-  return files.reduce((allPosts, postSlug) => {
+  const posts = files.reduce((allPosts, postSlug) => {
     const source = fs.readFileSync(path.join(postDir, locale, postSlug), 'utf8')
     const { data } = matter(source)
     return [{ ...data, slug: postSlug.replace('.mdx', '') }, ...allPosts]
   }, [])
+
+  // Generating RSS
+  RSS({ posts })
+
+  if (sort) {
+    return posts.sort((a, b) => (a.publishedAt < b.publishedAt ? 1 : -1))
+  }
+  return posts
 }
