+++ src/scripts/generate-rss.ts
@@ -11,9 +11,8 @@ const feed = new Feed({
   description,
   id: url,
   link: url,
-  language: 'en',
   favicon: `${url}/favicon.ico`,
-  image: `${url}/assets/resources/profile.png`,
+  image: `${url}/assets/examples/profile.png`,
   copyright: `All rights reserved ${new Date().getFullYear()}, Kalwabed Rizki`,
   generator: `${url} RSS API`,
   feed: `${url}/rss.xml`,
@@ -24,7 +23,7 @@ const feed = new Feed({
   },
   author: {
     email,
-    link: siteConfig.socials.Github,
+    link: siteConfig.socials.Twitter,
     name: title
   }
 })
@@ -45,7 +44,7 @@ const MyFeed = ({ posts }: { posts: Post[] }) => {
       ],
       copyright: `All rights reserved ${new Date().getFullYear()}, Kalwabed Rizki`,
       link: `${url}/blog/${post.slug}`,
-      image: `${url}/assets/uploads${post.featuredImage.image}`,
+      image: `${url}${post.featuredImage.image}`,
       id: `${url}/blog/${post.slug}`
     })
   )
